requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
frozenlist==1.8.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
magic-filter==1.0.12
multidict==6.7.0
openai==2.3.0
propcache==0.4.1
protobuf==6.32.1
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.44
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
yandex-speechkit==1.5.0
yarl==1.22.0
